requirements.txt
ed25519
python-bitcoinlib
